Separate network types to its own file
